Move helper to context
